Query By ID

+++ src/main/java/com/project/filmeslist/repository/FilmsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.project.filmeslist.entities.Films;
 
-public interface FilmsRepository extends JpaRepository<Films, Integer> {
+public interface FilmsRepository extends JpaRepository<Films, Long> {
 	
 }
